Factor out the lock contention reporting facility implementation
