+++ crypto/build.info
@@ -97,10 +97,10 @@ SOURCE[../providers/libfips.a]=$CORE_COMMON
 # Central utilities
 $UTIL_COMMON=\
         cryptlib.c params.c params_from_text.c bsearch.c ex_data.c o_str.c \
-        threads_pthread.c threads_win.c threads_none.c threads_common.c \
-        initthread.c context.c sparse_array.c asn1_dsa.c packet.c \
-        param_build.c param_build_set.c der_writer.c threads_lib.c \
-        params_dup.c time.c
+        threads_lock_contention.c threads_pthread.c threads_win.c \
+        threads_none.c threads_common.c initthread.c context.c sparse_array.c \
+        asn1_dsa.c packet.c param_build.c param_build_set.c der_writer.c \
+        threads_lib.c params_dup.c time.c
 
 SOURCE[../libcrypto]=$UTIL_COMMON \
         mem.c mem_sec.c \
